[frontend] refactor indicatorTypesField + schema validator + update trads
