+++ opencti-platform/opencti-front/lang/front/fr.json
@@ -2645,7 +2645,7 @@
   "Impact": "Impact",
   "indicators currently impacted by this rule": "indicateurs actuellement affectés par cette règle",
   "Life curve": "Courbe de vie",
-  "Apply on indicator observable types": "Appliquer sur les types d'observable des indicateurs",
+  "Apply on indicator observable types (none = ALL)": "Appliquer sur les types d'observable des indicateurs (aucun = TOUS)",
   "Define at which score thresholds the indicator is updated.": "Définir les seuils de score à partir desquels l'indicateur est mis à jour.",
   "Add a reaction point": "Ajouter un point de réaction",
   "Remove this reaction point": "Supprimer ce point de réaction",
